Update 04.markdown
The comment mentioning showing the TLL value says "// after 13s" - however, after 113 seconds it would show 7 seconds. So either that output value should be 7 (if the command was input after 113 seconds), or if it shows an output of 113 it should be "// after 7s" - I've made an edit that would change the comment and not the code and also added another comment after the value has expired to be more robust.